third_party/blink/web_tests/fast/dom/custom/unresolved-pseudoclass-expected.txt

CONSOLE WARNING: :unresolved pseudo selector is deprecated and will be removed in M80, around February 2020. Please use :not(:defined) instead. See https://www.chromestatus.com/features/4642138092470272 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
CONSOLE WARNING: line 34: document.registerElement is deprecated and will be removed in M80, around February 2020. Please use window.customElements.define instead. See https://www.chromestatus.com/features/4642138092470272 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Tests the :unresolved pseudoclass.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.querySelector("x-x:unresolved") is a
PASS window.getComputedStyle(a).color is "rgb(0, 0, 222)"
PASS window.getComputedStyle(b).color is "rgb(0, 0, 222)"
PASS window.getComputedStyle(b).borderColor is "rgb(0, 0, 111)"
PASS window.getComputedStyle(c).color is "rgb(0, 222, 0)"
PASS window.getComputedStyle(a).color is "rgb(0, 222, 0)"
PASS window.getComputedStyle(d).color is "rgb(0, 111, 0)"
PASS window.getComputedStyle(b).color is "rgb(0, 111, 0)"
PASS successfullyParsed is true

TEST COMPLETE
